1.20.1 initial release
Still in progress but fully functional!

+++ src/main/java/net/ddraig/suprememiningdimension/procedures/CavernsPlayerEntersDimensionProcedure.java
@@ -30,11 +30,11 @@ public class CavernsPlayerEntersDimensionProcedure {
 					if (world instanceof ServerLevel _serverworld) {
 						StructureTemplate template = _serverworld.getStructureManager().getOrCreate(new ResourceLocation("supreme_mining_dimension", "goldport"));
 						if (template != null) {
-							template.placeInWorld(_serverworld, new BlockPos(x, 64, z), new BlockPos(x, 64, z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
+							template.placeInWorld(_serverworld, new BlockPos((int)x, 64, (int)z), new BlockPos((int)x, 64, (int)z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
 						}
 					}
 					if (world instanceof Level _level)
-						CavernsPortalBlock.portalSpawn(_level, new BlockPos(x, 63, z));
+						CavernsPortalBlock.portalSpawn(_level, new BlockPos((int)x, 63, (int)z));
 					{
 						Entity _ent = entity;
 						_ent.teleportTo(x, 66, z);
@@ -56,17 +56,17 @@ public class CavernsPlayerEntersDimensionProcedure {
 			if (world instanceof ServerLevel _serverworld) {
 				StructureTemplate template = _serverworld.getStructureManager().getOrCreate(new ResourceLocation("supreme_mining_dimension", "airey"));
 				if (template != null) {
-					template.placeInWorld(_serverworld, new BlockPos(x, 124, z), new BlockPos(x, 124, z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
+					template.placeInWorld(_serverworld, new BlockPos((int)x, 124, (int)z), new BlockPos((int)x, 124, (int)z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
 				}
 			}
 			if (world instanceof ServerLevel _serverworld) {
 				StructureTemplate template = _serverworld.getStructureManager().getOrCreate(new ResourceLocation("supreme_mining_dimension", "goldport"));
 				if (template != null) {
-					template.placeInWorld(_serverworld, new BlockPos(x, 124, z), new BlockPos(x, 124, z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
+					template.placeInWorld(_serverworld, new BlockPos((int)x, 124, (int)z), new BlockPos((int)x, 124, (int)z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
 				}
 			}
 			if (world instanceof Level _level)
-				CavernsPortalBlock.portalSpawn(_level, new BlockPos(x, 64, z));
+				CavernsPortalBlock.portalSpawn(_level, new BlockPos((int)x, 64, (int)z));
 			{
 				boolean _setval = false;
 				entity.getCapability(SupremeMiningDimensionModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
